Improvements and bug fixes
- The browser texture is now created on the render thread.
- Browser pages now repaint after alt-tab.
- Browser dropdown menus are no longer drawn in the corner.
- A browser that starts slowly is now shown and closed properly.
- Browser clicks, drags and mouse releases now work correctly.
- Shift+A no longer types "AA" in the browser.
- Fast key presses are no longer missed.
- High-resolution scroll wheels now scroll the browser.
- Text is no longer drawn under shapes queued before it.
- The default font now loads, and the fallback no longer exits the game.
- System fonts are no longer duplicated on every vid_restart.
- D3D9 hooks are no longer removed and reinstalled on vid_restart, which could crash with overlays.
- Reading game memory no longer makes pages briefly read-only.
- A failing background task or HTTP callback no longer kills the game.
- Quitting no longer hangs on stalled HTTP requests.
- HTTP requests are limited to http and https.
- Zip compression now reports write failures, so asset dumps aren't deleted.
- Zip extraction now fails on any bad entry and streams large files.
- Zip buffers are now freed correctly.
- Zip entries can't escape a . or empty target folder.
- Crash report cleanup can no longer crash at startup.
- The crash reporter keeps the dump if it can't save a copy.
- Escape while typing no longer closes the crash reporter.
- Large crash dumps no longer time out on upload.
- Shaders are no longer recompiled on every alt-tab.
- Shader compile warnings no longer leak memory.
- The key-bind widget no longer binds the click or Escape used to cancel.
- The debug memory editor now writes the right bytes.
- Saved layouts that fail to load fall back to defaults.
- Signature is now safe to use in an if.
- The engine's vcpkg.json now includes sentry-native.

+++ src/Engine/Core/Input/Mouse.hpp
@@ -9,7 +9,7 @@ namespace IzEngine
 		static inline vec2 Position;
 		static inline vec2 Delta;
 		static inline vec2 LastDelta;
-		static inline int ScrollDelta = 0;
+		static inline float ScrollDelta = 0;
 
 		static void Register();
 		static void Process(uint64_t msg, uint64_t state);
